requirements.txt
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0
scikit-image>=0.21.0
ffmpeg-python>=0.2.0
av>=10.0.0
OpenColorIO>=2.2.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0.0
pillow>=9.0.0